+++ arch/efm32/drivers/usart/usart_printk.c
@@ -46,7 +46,7 @@ static inline void printk_out_char(char c)
 
 static PRINTK_HANDLER(efm32_usart_printk_out)
 {
-  uint_fast8_t i;
+  size_t i;
 
   for (i = 0; i < len; i++)
   {
